+++ HISTORY.md
@@ -2,9 +2,10 @@ History
 ===============
 
 ### By now 2014-03-14
-* 2 major enhancement
+* 3 major enhancement
 	* Completed the design of architectures
 	* Added BioTCM::Modules::Utility
+	* Added BioTCM::Modules::WorkingDir
 
 ### 0.0.0 2014-03-06
 * 1 major enhancement